requirements.txt
fastapi
uvicorn
pydantic
requests
openai
python-dotenv
pinecone-client
pandas
tqdm
numpy
